requirements.txt
flask
flask-sqlalchemy
openai
python-dotenv
numpy
scikit-learn
google-generativeai